Added more unit tests and switched to pytest
