+++ startup/requirements.txt
@@ -3,3 +3,4 @@ matplotlib==3.9.2
 Pillow==10.4.0
 PyDrive==1.3.1
 spotipy==2.24.0
+pytest==7.4.0
